+++ pyanalyze/functions.py
@@ -37,6 +37,7 @@ from .value import (
     SubclassValue,
     TypedValue,
     TypeVarValue,
+    make_mutable,
     unite_values,
     UnpackedValue,
     Value,
@@ -341,6 +342,7 @@ def compute_parameters(
             else:
                 # normal method
                 value = enclosing_class
+            value = make_mutable(value)
         else:
             # This is meant to exclude methods in nested classes. It's a bit too
             # conservative for cases such as a function nested in a method nested in a